requirements_ui.txt
streamlit>=1.35.0
chromadb==0.5.3
sentence-transformers==3.0.1
google-genai>=0.8.0
rouge-score==0.1.2
pytest==8.2.2
pandas>=2.0.0
